[aarch64] Add support for AArch64 Advanced SIMD.
The following patch enables SLEEF to use the Advanced SIMD vector
extension for the AArch64 execution mode of the A64 instruction set on
ARMv8 (AArch64 AdvSIMD hereafter).

The suffix "advsimd" is consistently used across the sources and the
build and test systems to identify this extension.

The patch have been tested on native hardware for single and double
precision for the following cases:

* iut: scalar
* iutadvsimd: AArch64 AdvSIMD
* iutcevext: clang vector extentions on AArch64

The patch includes also the following minor changes:

* the copyright notice in the README file has been changed to assign
  copyright to other contributors;
* libsleef-dft has been temporarly disabled for AArch64 builds;
* set some shell scripts permission to executable.