This is a dummy commmit for adding explanation of a dispatcher to commit messages.
Dispatcher is a mechanism for the library to automatically detect the available features of the CPU and use the fastest implementation from the functions with the same specification.
The patch adds the dispatcher for 128-bit computation that selects from SSE2, SSE4 and AVX2 implementations,
and the dispatcher for 256-bit computation that selects from AVX, FMA4 and AVX2.