[sinpi] Fix bug in sinpi series of functions
This patch fixes a bug that was making the error of sinpi larger than the designed limit.
The following functions have the same bug : sinpi, sinpif, cospi, cospif, sincospi, sincospif.
Besides, tgamma and tgammaf are affected by this bug.

This patch also adds regression tests for this bug.
Additionally, this patch adds modification to the tester that prevents false detection of error in hypot_u35.