[Makefile] Prevent build failure when build in parallel
With this patch, build failure in a parallel build is prevented.
Sometimes build failed in a parallel build since sleef.h is read by another process before it is written completely.
With this patch, sleef.h is once output to sleef.h.0 and then renmaed to sleef.h.
If mv command is atomic, build failure should be prevented.