* [gnuabi] Enable "finite" names for glibc compatibility.
glibc/libm defines `finite` names for some of the C99 functions. This patch adds such names to `libsleefgnuabi.so`.

The finite names are added only if `__attribute__((weak, alias("finite_name")))` is supported.

A set of tests for linking against the finite names is added. The tests are currently linking against the finite `asin/acos/asinf/acosf` for SSE, AVX, AVX2 and AVX512F. Further testing covering all functions and vector targets will be added in a subsequent commit.