This patch improves computation speed of log1p.
Instead of calculating log1p with the generic log subroutine, the new functions calculate it using a new algorithm, which is a modification of generic log.
The ratio of computation time between the old and new log1p is 1.3.
This patch also includes some cosmetic changes and fixes of typos (ENABLE_FMA_DP was wrongly used in sleefsimdsp.c, and now replaced with ENABLE_FMA_SP).