Running tests on intel sde (#131)
This patch makes the testers to run on Intel SDE.
The SDE is only used when the corresponding extension is not available on the computer.

In this patch, the SDE package is downloaded from the URL retained in SDE_URL.
SDE_URL is a Travis secure environment variable. Since redistribution of SDE is not allowed, we need to hide where we are downloading the SDE package from.

Travis terminates running programs if it does not receive anything from stdout of the program for 10 minutes. So --verbose option is used in testing.

https://github.com/shibatch/sleef/issues/85