[Travis] aarch64 and armhf builds on travis (#159)
With this patch, the following functionalities are added.

* Cross compilation without qemu is made possible. This is basically implementation in #157.
* aarch64 and armhf builds and testing on travis are implemented.
* Configuration file for travis-ci is now cleaned up and they are separated to shell scripts. This provides greater control over what to do for each configuration.
* Tester now supports flush-to-zero mode regardless of the target architecture.

For aarch64 and armhf builds and testing on travis, docker has to be used. This is because only gcc-4.8 can be used without docker, and I couldn't find newer cross compilers.
Usage of sde is now turned off for linux-clang setting, since it takes long time. It is enabled only for x86_64-gcc setting only.

Before cross-compilation, native build has to be executed.

If qemu with binfmt_misc is available, native build can be skipped.
In this case, FALSE has to be set to CMAKE_CROSSCOMPILING variable in the toolchain file.
An emulator can be specified for testing.

cmake -DCMAKE_TOOLCHAIN_FILE=../../toolchain-aarch64-binfmt.cmake -DEMULATOR=qemu-aarch64-static ..