[gnuabi]  Compatibility tests (#162)
This set of tests make sure tht the symbols we generate in `libsleefgnuabi.so` are compatible with the naming conventions of `libmvec` in `glibc`.

Note: we ship log2[f] in the library but we don't provide a finite symbol for it, although GNU libm on x86 expects it. This will be fixed in a separate commit/PR.This issue is tracked in https://github.com/shibatch/sleef/issues/164.