[Benchmark] New benchmarking tool with gnuplot(2) (#168)
This patch replaces the old benchmarking tool with a new tool with gnuplot.

With this new tool, the graph showing execution time can be automatically drawn.
It is now easy to see the difference between different library versions or compilers.
This tool is used to draw the graphs on sleef.org web site.

This patch implements the tool described in https://github.com/shibatch/sleef/issues/9.

A makefile is used In this patch. It is used to compile the necessary tools.
This makefile is used as a script for measurement. It is handy since ProcData.java has to be also compiled.