[SVE] Add SVE support to DFT (#182)
This patch adds SVE support to the DFT library.

The DFT library strongly depends on the vector length, and thus it does not mix well with vector-length agnosticism. In this patch, the utilized vector length can be specified by CONFIG macro in helpersve.h. The DFT subroutine uses a dispatcher to choose the kernel subroutine from those compiled for 2048, 1024, 512 and 256-bit vector lengths.

* CHANGELOG.md is updated according to the changes made.
* An error will be generated if VECTLENDP or VECTLENSP are redefined at helpersve.h
* Added explanation of the meaning of VECTLENDP and VECTLENSP to helpersve.h.