Add quadprecision math library (#235)
This is a part of implementation of issue #233 ( https://github.com/shibatch/sleef/issues/233 ).
At this point, add, mul, div and sqrt with testers are implemented. Remaining functions will be committed in the succeeding PRs.
As for vector extensions, SSE2, AVX, FMA4, AVX2, AV2_128, AVX512F, AdvSIMD and SVE are supported.

This quad-precision math library is built only if -DBUILD_QUAD option is given to cmake. For some time(1 year?), this sub-project is positioned at alpha development stage.