[Quad] Add functions for conversion between quad and string (#237)
This patch adds Sleef_strtoq and Sleef_qtostr which can be used to convert between a quad value and a string. These functions are not vectorized. The corresponding testers are also added.

This patch also adds functions for subtraction.

Intel compiler testing is temporarily disabled because of license expiration( https://github.com/shibatch/sleef/issues/238 ).