malloc is in stdlib.h on OpenBSD (#274)
With this simple fix your library builds without errors on OpenBSD and all tests pass except for the ilogb denormal/nonnumber tests where "correct" return INT_MIN for inf arg. This appears to be due to a bug in the OpenBSD amd64 implementation of ilogb as the documentation states that the function should return INT_MAX for inf arg. Tests return INT_MAX.