Atomise sleef.h generation
Commands in the list of SLEEF_HEADER_COMMANDS were sometimes observed
to overwrite each other when doing a parallel build. Instead of
building a list of commands, build a list of temporary files to depend
on, add a custom_command for each of them, and cat them all together
at the end.