Atomise commands used to generate quad inline headers
Individual COMMANDS in define_custom_command are not guaranteed to
execute sequentially, meaning that the inline headers may be broken if
doing a parallel build with make. Each command is extracted into its
own custom_command (some are merged, such as subsequent sed calls) so
that CMake can properly make sense of the dependencies and generate
consistent headers in parallel.