Support parallel builds with make (#491)
Individual COMMANDS in define_custom_command are not guaranteed to
execute sequentially, meaning that the inline headers may be broken if
doing a parallel build with make. Each command is extracted into its
own custom_command (some are merged, such as subsequent sed calls) so
that CMake can properly make sense of the dependencies and generate
consistent headers in parallel.

* Atomise creation of inline header, alias, dispscalar, simd disp, sleefquad header and build
* Reuse existing atomic inline libm header generation for CUDA

Now that custom_commands are atomic, parallel builds should be
supported with make.

* Use CMake cat instead of shell cat

This requires minimum CMake 3.18, which is not particularly
recent. This provides a better cross-platform way of concatenating
files.