Add qemu command prefix to testervecabi tests
This test suite covers the usage of scalar symbols
with auto-vectorisation, to test the GNU ABI
symbols generated by SLEEF.
However, some failures are recorded when emulating
these tests using user mode QEMU, for a
cross-compile build for AArch64 and s390x.
This is because this test suite does not get called
with the EMULATOR command necessary.
This fix adds the emulator to the test command
which makes the tests pass when cross-compiling
and using user mode QEMU emulation.