Reconcile RVV mask types. (#520)
Make the vmask types for single-precision and double-precision
interfaces compatible, so that we have fewer divergent functions.

Keep the vopmask types incompatible, because the underlying vbool*_t
types still encode some underlying assumptions about the number of lanes
involved.

Make the still-incompatible function names distinct between SP and DP,
and then reconcile these with the preprocessor at the end of the file so
that the caller still sees the expected operations.  This will help when
it comes to generating inline headers in a future revision.