Clean up RVV register composition. (#521)
Prefer use of __riscv_vcreate_*() intrinsics, where appropriate.  This
does away with piecemeal assembly of a tuple register types and ensures
that every component is defined.  It also gets rid of uninitialised
variable warnings.

__riscv_vcreate intrinsics aren't available in older toolchains but will
be available in version 1.0 of the intrinsics interface.  In the interim
include substitute macros which try to fill the gap as best as possible.

Use a regular pattern of cast-then-insert, and extract-then-cast when
dealing with tuple registers with mixed component types.

Simplify two vcreate macros in helperrvv.h, remove a few unused macros, etc..