Fix installation of shared library symlink
This fixes the regression from 29391ccd9d513132fc802a198f8b801fb273cb59.

The dummy target that was removed in 29391ccd9d513132fc802a198f8b801fb273cb59
was responsible for installing the unversioned symlink for the shared libraries.
The only part of this that was causing issue appears to be the dummy destination
that was only added as an workaround for old cmake version (pre 3.14-ish).
With the cmake version requirement bumped to 3.18 this should not be an issue anymore,
and we could also use the new NAMELINK_COMPONENT option to avoid using two install
statements.

This also fixes a bug from eb3d97785cb4f6f6f4c0ffd6570ca6c73ccac2ec causing
the libsleefgnuabi library to be missing the symlink...