Add tests for examples and SLEEF as subproject
We only want to check if examples are working and documented
use of SLEEF as subproject is not broken.
No need to complicate with different compilers, ...
Add a new workflow to run examples.
Add a new workflow for nested project and submodule.
Only works for x86 runners for now.
Fix exit error code in dft tutorial.c
Add project names to top-level CMakeLists to remove CMake warning.