Port docs to md rendered to html by Github Pages.
Use https://just-the-docs.com/ template. And setup Home page.

Need to use custom deployment workflow to be able to use docs/
as source, which is slightly inconvenient compared to other themes
but might allow more customisation later.

The benefit of deploying docs is that the main/landing page will
display automatically when clicking on docs in Github. Then navigation
is straightforward.