Move sections to dedicated folders and convert landing page
Sections' landing pages are now rendered in github (section/README.md)
and on website (section/index.html).

Providing the pages permalink will make links to other pages
work in both Github and website.
Note the / matters at the end of the permalinks it allows to maintain
directory structure (in presence of parent/child sections) or not,

Fix tables of content.

Links to images and test files, and old html content will break,
it is fixed in following commits.