Enable GNUABI build on more targets #525
Previously SLEEF_BUILD_GNUABI_LIBS was silently ignored except for a
small number of targets, or if the compiler does not support weak
aliases.

It can now be built regardless of OS and compiler on AArch64 and
x86_64 - trying to enable it on any other target is now an error.

The GLIBC *_finite symbols are handled using the same workaround as is
used for the DALIAS macro when compiling on a target which does not
support aliases. Though in practice these symbols are unlikely to be
required on systems where aliases are unsupported, they are part of
the API so need adding, if just to make the tests build.

Also removed two CMake variables:
- ENABLE_ALIAS, which was unused
- ENABLE_GNUABI, which was now redundant but could still be forced on,
  leading to cryptic build failures
Both names are still used by the preprocessor in sleefsimd* sources
for managing names and aliases.