Add issue and PR templates.
Add a CONTRIBUTING.md and a section in the docs on contributing.
Rm CONTRIBUTORS.md because it is outdated, redundant with Github stats
and hard to maintain with the large amount of new contributors.

Kept the PR template simple to not demotivate people from contributing,
it could be improved with specialized templates in the future if
necessary.
In order to improve bug reports, we offer several templates for
filing issues/bugs.