Configure.cmake: improve and unify RISC-V vector extension checks
To test for the RISC-V vector extension (RVV), we currently run two
small programs -- one for each value of LMUL=1,2 -- that use some
vector intrinsics from the v0.11.x draft of the intrinsics spec. The
library itself however needs newer intrinsics than the ones that we
test for. This can lead to a problem because GCC 13 supports the
intrinsics that we test for, but not (all of) the ones we intend to
use. In short, the build can fail with GCC 13:

  https://github.com/shibatch/sleef/issues/579

This commit changes the detection mechanism. Instead of compiling a
program that makes use of the intrinsics, we now just check the two
values, __riscv_v and __riscv_v_intrinsic, to ensure that we have
version 1.0 of the vector extension and version 0.12 of the
intrinsics. These should be the stable versions of each.

Note: the "m1" and "m2" checks are now identical because they are both
covered by the official spec.

Fixes #579