Add libm benchmarks
Extended current benchmark tool with support to benchmark of libm
functions. Libm benchmark is enabled in all functions listed in
benchlibm.cpp.
It interfaces with the rest of the existing tool through
benchmark_callers_libm.hpp (defines macros which simplify how the
list of functions to benchmark looks like as it groups functions
by precision/accuracy/variant).
The tool is integrated with SLEEF via CMake, meaning it can
be built automatically when SLEEF is built. To enable that,
pass CMake argument -DSLEEF_BUILD_BENCH=ON and
-DSLEEF_BUILD_BENCH_REF=ON.
Note this option is only enabled on Linux OS.